+++ es6/docxtemplater.js
@@ -26,7 +26,7 @@ const {
 	throwApiVersionError,
 } = require("./errors");
 
-const currentModuleApiVersion = [3, 23, 0];
+const currentModuleApiVersion = [3, 24, 0];
 
 const Docxtemplater = class Docxtemplater {
 	constructor(zip, { modules = [], ...options } = {}) {